Renamed pkg/articulation to pkg/abi
