+++ src/apps/chifra/pkg/abi/load_abi.go
@@ -1,4 +1,4 @@
-package articulation
+package abi
 
 import (
 	"encoding/json"
