+++ src/apps/chifra/pkg/abi/load_abi_test.go
@@ -1,4 +1,4 @@
-package articulation
+package abi
 
 import (
 	"testing"
