--- src/apps/chifra/pkg/articulation/load_abi.test.go
@@ -1,1 +1,0 @@
-package articulation
